Moved test methods from SampleTest to SanityTest
